Add files via upload
The objective of this application is to provide a centralized platform that enhances campus safety and 
accessibility to essential services. It aims to enable users to report safety concerns, maintenance issues, 
and request services such as counseling, medical assistance, or facility repairs in a quick and efficient 
manner. The app also ensures easy access to emergency contacts and institutional support resources, 
fostering a secure, supportive, and well-maintained campus environment through streamlined 
communication and timely assistance, user-friendly platform to ensure timely reporting, quick service 
access, and improved campus safety and well-being and Enhancing campus safety and services.